exporter/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in
aiohttp==3.9.5
    # via
    #   datasets
    #   fsspec
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
attrs==23.2.0
    # via aiohttp
certifi==2024.6.2
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   typer
    #   uvicorn
datasets==2.20.0
    # via -r requirements.in
dill==0.3.8
    # via
    #   datasets
    #   multiprocess
dnspython==2.6.1
    # via email-validator
email-validator==2.2.0
    # via fastapi
fastapi==0.111.0
    # via -r requirements.in
fastapi-cli==0.0.4
    # via fastapi
filelock==3.15.4
    # via
    #   datasets
    #   huggingface-hub
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.5.0
    # via
    #   datasets
    #   huggingface-hub
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via fastapi
huggingface-hub==0.23.4
    # via datasets
idna==3.7
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   yarl
jinja2==3.1.4
    # via fastapi
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.16
    # via datasets
numpy==2.0.0
    # via
    #   datasets
    #   pandas
    #   pyarrow
orjson==3.10.5
    # via fastapi
packaging==24.1
    # via
    #   datasets
    #   huggingface-hub
pandas==2.2.2
    # via datasets
prometheus-client==0.20.0
    # via -r requirements.in
psycopg2-binary==2.9.9
    # via -r requirements.in
pyarrow==16.1.0
    # via datasets
pyarrow-hotfix==0.6
    # via datasets
pydantic==2.7.4
    # via fastapi
pydantic-core==2.18.4
    # via pydantic
pygments==2.18.0
    # via rich
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   uvicorn
python-multipart==0.0.9
    # via fastapi
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via
    #   datasets
    #   huggingface-hub
    #   uvicorn
requests==2.32.3
    # via
    #   datasets
    #   huggingface-hub
rich==13.7.1
    # via typer
shellingham==1.5.4
    # via typer
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.30
    # via -r requirements.in
starlette==0.37.2
    # via fastapi
tqdm==4.66.4
    # via
    #   datasets
    #   huggingface-hub
typer==0.12.3
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   fastapi
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer
tzdata==2024.1
    # via pandas
ujson==5.10.0
    # via fastapi
urllib3==2.2.2
    # via requests
uvicorn==0.30.1
    # via fastapi
uvloop==0.19.0
    # via uvicorn
watchfiles==0.22.0
    # via uvicorn
websockets==12.0
    # via uvicorn
xxhash==3.4.1
    # via datasets
yarl==1.9.4
    # via aiohttp
